Add files via upload
This is the VBA script I used and a screenshot of the results. Every time I tried to run the script on the larger data set my computer crashed, but it worked on the smaller data set so the results will be consistent across both. The VBA script is wrote to be used with the larger multiple year data set.